BM-23 fix rebase
